requirements.txt
nest_asyncio
streamlit
duckduckgo_search
phi.tools